refactor: Split OAuth2ConfigBuilder into builder.rs, move OAuth2 config & client under oauth2/config folder.
