+++ src/builder.rs
@@ -44,8 +44,8 @@ use tokio::sync::{Notify, RwLock};
 
 use crate::constant::DEFAULT_ESI_URL;
 use crate::error::EsiError;
-use crate::oauth2::client::OAuth2Client;
-use crate::oauth2::config::OAuth2Config;
+use crate::oauth2::config::client::OAuth2Client;
+use crate::oauth2::OAuth2Config;
 use crate::EsiClient;
 
 /// Builder for configuring and constructing an `EsiClient`.
